requirements.txt
# Core ML + audio
tensorflow-cpu==2.15.0
tensorflow_hub==0.16.1
numpy>=1.24
librosa==0.10.1
resampy>=0.4.2
sounddevice>=0.4.6

# Playback + config
pygame>=2.5.2
PyYAML>=6.0.1
